requirements.txt
wikiciteparser
pywikibot
requests
docopt
unidecode
subprocess32
pypdf2
flask
pyyaml
mwoauth
requests_oauthlib
